+++ doc/version_history.md
@@ -1,5 +1,20 @@
 # Version History
 
+0.1.1
+
+- Add the dependencies of `crc`, `approx`, and `ts_control_utils`.
+- Add the daq classes: **Flowmeter**, **PowerGridMonitor**, and **PierFan**.
+- Add the mock classes: **MockFlowmeterGroup**, **MockPressureTransducerGroup**, **MockPowerGridMonitor**, and **MockPierFan**.
+- Add the `utility.rs` and `enums.rs`.
+- Update the `constants.rs` and `mock_constants.rs`.
+- Remove the **Default** in **MockPressureTransducer**.
+- Use the `f32` instead of `f64` in **MockPressureTransducer** and **MockTemperatureHub**.
+- Add the `parameters_app.yaml`.
+- Add the **Config**.
+- Update the **MockPlant**.
+- Update the class diagram.
+- Update the **README.md**.
+
 0.1.0
 
 - Initial version to setup the repository.